restructure jointDemand: get rid of unnecessary subpackage
