+++ contribs/commercialTrafficApplications/src/test/java/org/matsim/contrib/commercialTrafficApplications/jointDemand/DefaultCommercialJobGeneratorTest.java
@@ -1,9 +1,10 @@
-package org.matsim.contrib.commercialTrafficApplications.jointDemand.commercialJob;
+package org.matsim.contrib.commercialTrafficApplications.jointDemand;
 
 import org.junit.Ignore;
 import org.junit.Rule;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.TransportMode;
+import org.matsim.contrib.commercialTrafficApplications.jointDemand.DefaultCommercialJobGenerator;
 import org.matsim.contrib.commercialTrafficApplications.jointDemand.TestScenarioGeneration;
 import org.matsim.contrib.freight.carrier.CarrierVehicleTypeWriter;
 import org.matsim.contrib.freight.carrier.CarrierVehicleTypes;
